+++ CHANGELOG.rst
@@ -5,8 +5,16 @@ v (next)
 --------
 
 
+v21.6.2
+-------
+
+- Add new --list-formats command line option to list supported archive formats
+- Add new exttractcode.api.extract_archive() API function to extract a single
+  archive file of any supported format, non recursively.
+
+
 v21.6.1
---------
+-------
 
 - Add support for VMDK, QCOW and VDI VM image filesystems extraction
 - Add new configuration mechanism to get third-party binary paths:
@@ -20,13 +28,13 @@ v21.6.1
 - Update to use latest skeleton
 
 
-v2021-2-24
+v21-2-24
 ----------
 
 - Fix incorrect documentation link
 
 
-v2021-1-21
+v21-1-21
 ----------
 
 - Fix bug related to CommonCode libraries loading
@@ -36,7 +44,7 @@ v2021-1-21
 - Reorganize tests files
 
 
-v2021-1-15
+v21-1-15
 ----------
 
 - Drop support for Python 2